+++ litex/soc/integration/builder.py
@@ -43,7 +43,9 @@ soc_software_packages = [
     "libcompiler_rt",
 
     # LiteX cores.
-    "libbase",
+    "libutils",
+    "libcomm",
+    "libmisc",
 
     # LiteX Ecosystem cores.
     "libfatfs",
