+++ litex/soc/software/bios/cmds/cmd_mem.c
@@ -2,7 +2,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <base/memtest.h>
+#include <libutils/memtest.h>
 
 #include <generated/csr.h>
 #include <generated/mem.h>
