+++ litex/soc/software/bios/isr.c
@@ -7,7 +7,7 @@
 #include <generated/csr.h>
 #include <generated/soc.h>
 #include <irq.h>
-#include <base/uart.h>
+#include <libcomm/uart.h>
 #include <stdio.h>
 
 #if defined(__microwatt__)
