+++ litex/soc/software/libcomm/spiflash.h
