+++ litex/soc/software/libmisc/sim_debug.c
@@ -1,4 +1,4 @@
-#include <base/sim_debug.h>
+#include "sim_debug.h"
 
 #include <stdio.h>
 #include <generated/csr.h>
