+++ litex/soc/software/libmisc/sim_debug.h
