+++ litex/soc/software/libutils/console.h
