+++ litex/soc/software/libutils/crc.h
