+++ litex/soc/software/libutils/crc16.c
@@ -1,4 +1,4 @@
-#include <base/crc.h>
+#include "crc.h"
 
 #ifndef SMALL_CRC
 static const unsigned int crc16_table[256] = {
