+++ litex/soc/software/libutils/lfsr.h
