+++ litex/soc/software/libutils/progress.h
