+++ litex/soc/software/libutils/system.c
